Run drag and resize listener outside of ng zone
The dragging and reszie listener are running outside of the angular zone
now. This is supposed to be provide better performance since the run
asynchrone and we set the component manually ready for check which runs
change detection. But rather than run for the entire component tree
only the component and its children are marked for change detection.