WIP4: snapping functionality - remove position dependency of tasks
The tasks shared the same position object which was passed as input to
all of them. so changing it changed the position of all tasks. this is
fixed through reassigning the position in ngOnInit. However, i don't
quite understand why its working, since the object is not cloned but
just reassigned to another variable (which should pass the reference to
the same object in my understanding).