WIP: fix different size of cells in label-container and cell-container
One cell is supposed to have a total height of 30px and a label-cell
60px. Therefore twice the number of cells exist. Somehow there was
space between the vertically aligned label-cells. This could be fixed
using align-content: flex-start which is responsible for aligning
the content on the secondary axis. flex-start ensures that there is
no space in the beginning and between the elements. So its the same as
justify-content but not for the main-axis.